+++ elixir/config/config.exs
@@ -4,6 +4,7 @@ import Config
 
 config :boj_rest,
   port: String.to_integer(System.get_env("BOJ_PORT") || "7700"),
+  bind_ip: System.get_env("BOJ_BIND_IP") || "127.0.0.1",
   cartridges_root:
     System.get_env("BOJ_CARTRIDGES_ROOT") ||
       Path.expand("../../cartridges", __DIR__),